+++ lib/src/framework/core/dependencies/helpers/hover_shared_preferences_helper.dart
@@ -1,6 +1,6 @@
 import 'package:shared_preferences/shared_preferences.dart';
 
-class SharedPreferencesHelper {
+class HoverSharedPreferencesHelper {
   static SharedPreferences _instance;
 
   static Future<SharedPreferences> getInstance() async {