+++ lib/src/framework/core/hover_page_base.dart
@@ -2,12 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import 'structure/hover_scaffold.dart';
+import 'structure/hover_snackbar_displayer.dart';
+import 'dependencies/helpers/hover_snackbar_helper.dart';
 import 'dependencies/navigation/routing/hover_route.dart';
 import 'dependencies/navigation/routing/hover_router.dart';
 import 'dependencies/global_widgets/hover_global_widgets.dart';
 
-abstract class HoverPageBase extends HoverScaffold implements HoverRoute, HoverNavigation {
-  static const Duration _defaultSnackBarDuration = Duration(seconds: 2);
+abstract class HoverPageBase extends HoverScaffold implements HoverRoute, HoverNavigation, HoverSnackBarDisplayer {
   final String title;
   final Color backgroundColor;
 
@@ -41,18 +42,6 @@ abstract class HoverPageBase extends HoverScaffold implements HoverRoute, HoverN
         floatingActionButton: buildFloatingActionButton(context),
       ),
     );
-  }
-
-  void showSnackBar(BuildContext context, String message, {Duration duration: _defaultSnackBarDuration}) {
-    //If the buildSnackBar method was overidden and doesn't return null,
-    //use the snackbar created from that method. Else, use globalWidgets.snackBar.
-    final globalWidgets = getGlobalWidgets(context);
-    var snackBar = buildSnackBar(context);
-    snackBar = (snackBar != null) ? snackBar : globalWidgets.snackBarBuilder(context);
-
-    //If both the builSnackBar() and globalWidgets.snackBar are null, use a basic snackbar.
-    final _snackBar = (snackBar != null) ? snackBar : SnackBar(content: Text(message), duration: duration);
-    Scaffold.of(context).showSnackBar(_snackBar);
   }
 
   HoverRoutingManager _getAppNavigationManager(BuildContext context) {
@@ -97,15 +86,26 @@ abstract class HoverPageBase extends HoverScaffold implements HoverRoute, HoverN
     }
   }
 
-  Widget buildSnackBar(BuildContext context) {
-    return getGlobalWidgets(context).snackBarBuilder(context);
-  }
-
   Widget buildFloatingActionButton(BuildContext context) {
     if (getGlobalWidgets(context).floatingActionButtonBuilder != null) {
       return getGlobalWidgets(context).floatingActionButtonBuilder(context);
     } else {
       return SizedBox.shrink();
     }
+  }
+
+  @override
+  void showSnackBar(BuildContext context, Widget content, {Duration duration}) {
+    HoverSnackBarHelper.showSnackBar(context, content, duration: duration);
+  }
+
+  @override
+  void showPlainSnackBar(BuildContext context, String message, {Duration duration}) {
+    HoverSnackBarHelper.showPlainSnackBar(context, message, duration: duration);
+  }
+
+  @override
+  Widget regenerate(BuildContext context) {
+    return build(context);
   }
 }