+++ lib/src/framework/core/structure/hover_snackbar_displayer.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+
+abstract class HoverSnackBarDisplayer {
+  void showSnackBar(BuildContext context, Widget content, {Duration duration});
+  void showPlainSnackBar(BuildContext context, String message, {Duration duration});
+}